Created a register controller
