+++ src/models/user.js
@@ -25,5 +25,5 @@ const userSchema = new Schema({
     }
 });
 
-const User = mongoose.model('User', userSchema);
-module.exports = User;
+const user = mongoose.model('user', userSchema);
+module.exports = user;
